Added fix for zero decimals that are not zero.
Changed tax calculation to be based on tax rate instead of cart values.

Changed how we calculate the order total, basing it onthe klarna order lines instead of cart. Then check this against the cart total to get a difference, that we then add as a seperate order line to the Klarna order, to get the same price in WooCommerce and Klarna.