Update the WC billing address on change
The "change" event is triggered when the customer modifies their billing address, however, we're not handling this in the WC hidden form.

Since WPML adapts the currency based on the billing country, the country, and by extension, the currency, will remain stale when the customer changes country in the Klarna form.