Use first country available as billing country for subscriptions
If the merchant set the default location to "No location by default", no billing country will be specified in the checkout page which effect customer's whose billing country is missing, and all guest customers. Usually, if the store only serves one country, WooCommerce automatically selects that country. However, the billing country won't be set until the customer provides their billing address. Given that there is only one option available to the customer, we can assume that it will be the selected country.