Skip state validation
There doesn't seem to be a standard naming convention for country states. E.g., In an Irish state in Woo is "D" (or "Dublin") whereas in Klarna it is "Co. Dublin". Furthermore, the way you enter the state in Woo is through a select dropdown whereas in Klarna it must be typed unless autocompleted.